pino: add a hook to record logging metrics
Pino by default reports log numbers. Since we need the logging level
labels, that's only on the logger object. From the config provided by
the Pino API, only hooks seem to bring in the logger context, from
which we can try to reverse the log level number to its label.

Closes #24.